pyproject.toml
[tool.poetry]
name = "translator-template"
version = "0.1.0"
description = ""
authors = ["Chris Thrasher <cthrash@github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]
pylint = "^2.8.3"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
rope = "^0.19.0"
black = "^21.5b2"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = 'translator_template.app:main'